+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to `zbar-php` will be documented in this file
 
 ## Unreleased
 
+- New `Zbar::type()` (to get the type of bar code) and `Zbar::decode()` (to get an object with both the bar code and type).
+
 ## 1.2.0 - 2021-06-03
 - PHP 8 support added.